Rename plotting to visualization according to #79 and consistent with #91 and #80
