+++ impedance/tests/test_visualization.py
@@ -1,6 +1,6 @@
 import matplotlib.pyplot as plt
 import numpy as np
-from impedance.plotting import plot_altair, plot_bode, plot_nyquist
+from impedance.visualization import plot_altair, plot_bode, plot_nyquist
 import json
